split download modal into separate sub-components
