+++ frontend/javascripts/admin/dataset/remote/add_remote_layer.tsx
@@ -11,7 +11,7 @@ import { useCallback, useEffect, useMemo, useState } from "react";
 import type { APIDataStore, VoxelSize } from "types/api_types";
 import type { DatasourceConfiguration } from "types/schemas/datasource.types";
 import type { ArbitraryObject } from "types/type_utils";
-import { Hint } from "viewer/view/action_bar/download_modal_view";
+import { Hint } from "viewer/view/action_bar/download_modal/download_info_shared";
 import { GoogleAuthFormItem } from "./google_auth_form_item";
 
 const FormItem = Form.Item;
