+++ frontend/javascripts/viewer/view/action_bar/download_modal/download_modal_view.tsx
@@ -1,68 +1,12 @@
-import { CopyOutlined } from "@ant-design/icons";
-import { useStartAndPollJob } from "admin/job/job_hooks";
-import {
-  doWithToken,
-  downloadAnnotation,
-  downloadWithFilename,
-  getAuthToken,
-  startExportTiffJob,
-} from "admin/rest_api";
-import {
-  Alert,
-  Button,
-  Checkbox,
-  Col,
-  Divider,
-  Modal,
-  Radio,
-  Row,
-  Tabs,
-  type TabsProps,
-  Tooltip,
-  Typography,
-} from "antd";
-import { LayerSelection } from "components/layer_selection";
-import features from "features";
-import { formatCountToDataAmountUnit, formatScale } from "libs/format_utils";
-import { makeComponentLazy, useFetch } from "libs/react_helpers";
-import { useWkSelector } from "libs/react_hooks";
-import Toast from "libs/toast";
-import {
-  computeArrayFromBoundingBox,
-  computeBoundingBoxFromBoundingBoxObject,
-  computeShapeFromBoundingBox,
-} from "libs/utils";
-import messages from "messages";
-import type React from "react";
-import { type ReactElement, useState } from "react";
-import {
-  type AdditionalAxis,
-  type APIDataLayer,
-  type APIDataset,
-  APIJobCommand,
-  type VoxelSize,
-} from "types/api_types";
-import type { BoundingBoxMinMaxType } from "types/bounding_box";
-import type { Vector3 } from "viewer/constants";
-import {
-  getByteCountFromLayer,
-  getDataLayers,
-  getLayerByName,
-  getMagInfo,
-} from "viewer/model/accessors/dataset_accessor";
-import { getAdditionalCoordinatesAsString } from "viewer/model/accessors/flycam_accessor";
-import { getUserBoundingBoxesFromState } from "viewer/model/accessors/tracing_accessor";
-import {
-  getReadableNameOfVolumeLayer,
-  getVolumeTracingById,
-  hasVolumeTracings,
-} from "viewer/model/accessors/volumetracing_accessor";
-import { Model } from "viewer/singletons";
-import type { StoreAnnotation, UserBoundingBox } from "viewer/store";
-import { MagSlider } from "viewer/view/action_bar/mag_slider";
-import { BoundingBoxSelection } from "viewer/view/ai_jobs/components/bounding_box_selection";
+import { Modal, Tabs, type TabsProps } from "antd";
 
-const { Paragraph, Text } = Typography;
+import { makeComponentLazy } from "libs/react_helpers";
+import { useWkSelector } from "libs/react_hooks";
+
+import { type ReactElement, useState } from "react";
+import { DownloadAnnotationTab } from "./download_annotation_tab";
+import { DownloadPythonTab } from "./download_python_tab";
+import { TiffExportTab } from "./download_tiff_export_tab";
 
 type TabKeys = "download" | "export" | "python";
 
@@ -74,204 +18,6 @@ type Props = {
   initialBoundingBoxId?: number;
 };
 
-type ExportLayerInfos = {
-  displayName: string;
-  layerName: string | null;
-  tracingId: string | null;
-  annotationId: string | null;
-  additionalAxes?: AdditionalAxis[] | null;
-};
-
-enum ExportFormat {
-  OME_TIFF = "OME_TIFF",
-  TIFF_STACK = "TIFF_STACK",
-}
-
-const EXPECTED_DOWNSAMPLING_FILE_SIZE_FACTOR = 1.33;
-
-const exportKey = (layerInfos: ExportLayerInfos, mag: Vector3) =>
-  `${layerInfos.layerName || ""}__${layerInfos.tracingId || ""}__${mag.join("-")}`;
-
-function getExportLayerInfos(
-  layer: APIDataLayer,
-  annotation: StoreAnnotation | null | undefined,
-): ExportLayerInfos {
-  const annotationId = annotation != null ? annotation.annotationId : null;
-
-  if (layer.category === "color" || !layer.tracingId) {
-    return {
-      displayName: layer.name,
-      layerName: layer.name,
-      tracingId: null,
-      annotationId: null,
-      additionalAxes: layer.additionalAxes,
-    };
-  }
-
-  // The layer is a volume tracing layer, since tracingId exists. Therefore, a tracing
-  // must exist.
-  if (annotation == null) {
-    // Satisfy TS.
-    throw new Error("Tracing is null, but layer.tracingId is defined.");
-  }
-  const readableVolumeLayerName = getReadableNameOfVolumeLayer(layer, annotation) || "Volume";
-  const volumeTracing = getVolumeTracingById(annotation, layer.tracingId);
-
-  return {
-    displayName: readableVolumeLayerName,
-    layerName: layer.fallbackLayerInfo?.name ?? null,
-    tracingId: volumeTracing.tracingId,
-    annotationId,
-    additionalAxes: layer.additionalAxes,
-  };
-}
-
-export function isBoundingBoxExportable(boundingBox: BoundingBoxMinMaxType, mag: Vector3) {
-  const shape = computeShapeFromBoundingBox(boundingBox);
-  const volume =
-    Math.ceil(shape[0] / mag[0]) * Math.ceil(shape[1] / mag[1]) * Math.ceil(shape[2] / mag[2]);
-  const volumeExceeded = volume > features().exportTiffMaxVolumeMVx * 1024 * 1024;
-  const edgeLengthExceeded = shape.some(
-    (length, index) => length / mag[index] > features().exportTiffMaxEdgeLengthVx,
-  );
-
-  const alerts = (
-    <>
-      {volumeExceeded && (
-        <Alert
-          type="error"
-          title={`The volume of the selected bounding box (${volume} vx) is too large. Tiff export is only supported for up to ${
-            features().exportTiffMaxVolumeMVx
-          } Megavoxels.`}
-        />
-      )}
-      {edgeLengthExceeded && (
-        <Alert
-          type="error"
-          title={`An edge length of the selected bounding box (${shape.join(
-            ", ",
-          )}) is too large. Tiff export is only supported for boxes with edges smaller than ${
-            features().exportTiffMaxEdgeLengthVx
-          } vx.`}
-        />
-      )}
-    </>
-  );
-
-  return {
-    isExportable: !volumeExceeded && !edgeLengthExceeded,
-    alerts,
-  };
-}
-
-function estimateFileSize(
-  selectedLayer: APIDataLayer,
-  mag: Vector3,
-  boundingBox: BoundingBoxMinMaxType,
-  exportFormat: ExportFormat,
-) {
-  const shape = computeShapeFromBoundingBox(boundingBox);
-  const volume =
-    Math.ceil(shape[0] / mag[0]) * Math.ceil(shape[1] / mag[1]) * Math.ceil(shape[2] / mag[2]);
-  return formatCountToDataAmountUnit(
-    volume *
-      getByteCountFromLayer(selectedLayer) *
-      (exportFormat === ExportFormat.OME_TIFF ? EXPECTED_DOWNSAMPLING_FILE_SIZE_FACTOR : 1),
-  );
-}
-
-function formatSelectedScale(dataset: APIDataset, mag: Vector3) {
-  const magAdaptedScale = dataset.dataSource.scale.factor.map((f, i) => f * mag[i]);
-  const unit = dataset.dataSource.scale.unit;
-  const scale = { factor: magAdaptedScale, unit } as VoxelSize;
-  return formatScale(scale);
-}
-
-export function Hint({
-  children,
-  style,
-}: {
-  children: React.ReactNode;
-  style: React.CSSProperties;
-}) {
-  return (
-    <div style={{ ...style, fontSize: 12, color: "var(--ant-color-text-secondary)" }}>
-      {children}
-    </div>
-  );
-}
-
-async function copyToClipboard(code: string) {
-  await navigator.clipboard.writeText(code);
-  Toast.success("Snippet copied to clipboard.");
-}
-
-function CopyableCodeSnippet({ code, onCopy }: { code: string; onCopy?: () => void }) {
-  return (
-    <pre>
-      <Button
-        style={{
-          float: "right",
-          border: "none",
-          width: "18px",
-          height: "16px",
-          background: "transparent",
-        }}
-        onClick={() => {
-          copyToClipboard(code);
-          if (onCopy) {
-            onCopy();
-          }
-        }}
-        icon={<CopyOutlined />}
-      />
-      {code}
-    </pre>
-  );
-}
-
-function getPythonAnnotationDownloadSnippet(authToken: string | null, annotation: StoreAnnotation) {
-  return `import webknossos as wk
-
-with wk.webknossos_context(
-    token="${authToken || "<insert token here>"}",
-    url="${window.location.origin}"
-):
-    annotation = wk.Annotation.download("${annotation.annotationId}")
-`;
-}
-
-function getPythonDatasetDownloadSnippet(authToken: string | null, dataset: APIDataset) {
-  const nonDefaultHost = !document.location.host.endsWith("webknossos.org");
-  const indentation = "\n        ";
-  const contextUrlAddendum = nonDefaultHost ? `, url="${window.location.origin}"` : "";
-  const maybeUrlParameter = nonDefaultHost
-    ? `${indentation}webknossos_url="${window.location.origin}"`
-    : "";
-
-  return `import webknossos as wk
-
-with wk.webknossos_context(token="${authToken || "<insert token here>"}"${contextUrlAddendum}):
-    # Download the dataset.
-    dataset = wk.Dataset.download(
-        dataset_name_or_url="${dataset.name}",
-        organization_id="${dataset.owningOrganization}",${maybeUrlParameter}
-    )
-    # Alternatively, directly open the dataset. Image data will be
-    # streamed when being accessed.
-    remote_dataset = wk.Dataset.open_remote(
-        dataset_name_or_url="${dataset.name}",
-        organization_id="${dataset.owningOrganization}",${maybeUrlParameter}
-    )
-`;
-}
-
-const okTextForTab = new Map<TabKeys, string | null>([
-  ["download", "Download"],
-  ["export", "Export"],
-  ["python", null],
-]);
-
 function _DownloadModalView({
   isOpen,
   onClose,
@@ -279,495 +25,79 @@ function _DownloadModalView({
   initialTab,
   initialBoundingBoxId,
 }: Props): ReactElement {
-  const activeUser = useWkSelector((state) => state.activeUser);
   const annotation = useWkSelector((state) => state.annotation);
   const dataset = useWkSelector((state) => state.dataset);
-  const rawUserBoundingBoxes = useWkSelector((state) => getUserBoundingBoxesFromState(state));
-  const currentAdditionalCoordinates = useWkSelector((state) => state.flycam.additionalCoordinates);
+
   const typeName = isAnnotation ? "annotation" : "dataset";
-  const isMergerModeEnabled = useWkSelector(
-    (state) => state.temporaryConfiguration.isMergerModeEnabled,
-  );
-  const hasVolumeFallback = annotation.volumes.some((volume) => volume.fallbackLayer != null);
-  const isVolumeNDimensional = annotation.volumes.some(
-    (tracing) => tracing.additionalAxes.length > 0,
-  );
-  const hasVolumes = hasVolumeTracings(annotation);
-  const initialFileFormatToDownload = hasVolumes ? "zarr3" : "nml";
 
   const [activeTabKey, setActiveTabKey] = useState<TabKeys>(initialTab ?? "download");
-  const [keepWindowOpen, setKeepWindowOpen] = useState(true);
-  const [fileFormatToDownload, setFileFormatToDownload] = useState<"zarr3" | "wkw" | "nml">(
-    initialFileFormatToDownload,
-  );
-  const [selectedLayerName, setSelectedLayerName] = useState<string>(
-    dataset.dataSource.dataLayers[0].name,
-  );
 
-  const layers = getDataLayers(dataset);
+  // const handleOk = async () => {
+  //   if (activeTabKey === "download") {
+  //     await Model.ensureSavedState();
+  //     const includeVolumeData = fileFormatToDownload === "wkw" || fileFormatToDownload === "zarr3";
+  //     downloadAnnotation(
+  //       annotation.annotationId,
+  //       annotation.annotationType,
+  //       hasVolumeFallback,
+  //       undefined,
+  //       fileFormatToDownload,
+  //       includeVolumeData,
+  //     );
+  //     onClose();
+  //   } else if (activeTabKey === "export" && startJob != null) {
+  //     await Model.ensureSavedState();
+  //     await startJob(async () => {
+  //       const job = await startExportTiffJob(
+  //         dataset.id,
+  //         computeArrayFromBoundingBox(selectedBoundingBox.boundingBox),
+  //         currentAdditionalCoordinates,
+  //         selectedLayerInfos.layerName,
+  //         mag.join("-"),
+  //         selectedLayerInfos.annotationId,
+  //         selectedLayerInfos.displayName,
+  //         exportFormat === ExportFormat.OME_TIFF,
+  //       );
+  //       return [exportKey(selectedLayerInfos, mag), job.id];
+  //     });
 
-  const selectedLayer = getLayerByName(dataset, selectedLayerName);
-  const selectedLayerInfos = getExportLayerInfos(selectedLayer, annotation);
-  const selectedLayerMagInfo = getMagInfo(selectedLayer.mags);
-
-  const userBoundingBoxes = [
-    ...rawUserBoundingBoxes,
-    {
-      id: -1,
-      name: "Full layer",
-      boundingBox: computeBoundingBoxFromBoundingBoxObject(selectedLayer.boundingBox),
-      color: [255, 255, 255],
-      isVisible: true,
-    } as UserBoundingBox,
-  ];
-
-  const [selectedBoundingBoxId, setSelectedBoundingBoxId] = useState(
-    initialBoundingBoxId ?? userBoundingBoxes[0].id,
-  );
-  const [rawMag, setMag] = useState<Vector3>(selectedLayerMagInfo.getFinestMag());
-  const mag = selectedLayerMagInfo.getClosestExistingMag(rawMag);
-  const [exportFormat, setExportFormat] = useState<ExportFormat>(ExportFormat.OME_TIFF);
-
-  const selectedBoundingBox = userBoundingBoxes.find(
-    (bbox) => bbox.id === selectedBoundingBoxId,
-  ) as UserBoundingBox;
-  const { isExportable, alerts: boundingBoxCompatibilityAlerts } = isBoundingBoxExportable(
-    selectedBoundingBox.boundingBox,
-    mag,
-  );
-
-  const { runningJobs: runningExportJobs, startJob } = useStartAndPollJob({
-    async onSuccess(job) {
-      if (job.resultLink != null) {
-        const token = await doWithToken(async (t) => t);
-        downloadWithFilename(`${job.resultLink}?token=${token}`);
-      }
-    },
-    onFailure() {
-      Toast.error("Error when exporting data. Please contact us for support.");
-    },
-  });
-
-  const handleOk = async () => {
-    if (activeTabKey === "download") {
-      await Model.ensureSavedState();
-      const includeVolumeData = fileFormatToDownload === "wkw" || fileFormatToDownload === "zarr3";
-      downloadAnnotation(
-        annotation.annotationId,
-        annotation.annotationType,
-        hasVolumeFallback,
-        undefined,
-        fileFormatToDownload,
-        includeVolumeData,
-      );
-      onClose();
-    } else if (activeTabKey === "export" && startJob != null) {
-      await Model.ensureSavedState();
-      await startJob(async () => {
-        const job = await startExportTiffJob(
-          dataset.id,
-          computeArrayFromBoundingBox(selectedBoundingBox.boundingBox),
-          currentAdditionalCoordinates,
-          selectedLayerInfos.layerName,
-          mag.join("-"),
-          selectedLayerInfos.annotationId,
-          selectedLayerInfos.displayName,
-          exportFormat === ExportFormat.OME_TIFF,
-        );
-        return [exportKey(selectedLayerInfos, mag), job.id];
-      });
-
-      if (!keepWindowOpen) {
-        onClose();
-      }
-    }
-  };
-
-  const maybeShowWarning = () => {
-    const volumeFallbackWarning =
-      activeTabKey === "download" && hasVolumeFallback ? (
-        <Row key="no-fallback">
-          <Text
-            style={{
-              margin: "0 6px 12px",
-            }}
-            type="warning"
-          >
-            {messages["annotation.no_fallback_data_included"]}
-          </Text>
-        </Row>
-      ) : null;
-    const pythonTokenWarning =
-      activeTabKey === "python" ? (
-        <Row key="python-token-warning">
-          <Text
-            style={{
-              margin: "0 6px 12px",
-            }}
-            type="warning"
-          >
-            {activeUser != null
-              ? messages["download.python_do_not_share"]({ typeName })
-              : messages["annotation.register_for_token"]}
-          </Text>
-        </Row>
-      ) : null;
-
-    return [volumeFallbackWarning, pythonTokenWarning];
-  };
+  //     if (!keepWindowOpen) {
+  //       onClose();
+  //     }
+  //   }
+  // };
 
   const handleTabChange = (key: string) => {
     setActiveTabKey(key as TabKeys);
   };
 
-  const handleKeepWindowOpenChecked = (e: any) => {
-    setKeepWindowOpen(e.target.checked);
-  };
-
-  const typeDependentFileName = isAnnotation ? "annotation files" : "datasets";
-  const moreInfoHint = (
-    <Hint
-      style={{
-        margin: "0px 12px 0px 12px",
-      }}
-    >
-      For more information on how to work with {typeDependentFileName} visit the{" "}
-      <a
-        href="https://docs.webknossos.org/webknossos/data/export_ui.html"
-        target="_blank"
-        rel="noreferrer"
-      >
-        user documentation
-      </a>
-      .
-    </Hint>
-  );
-
-  const workerInfo = (
-    <Row>
-      <Divider
-        style={{
-          margin: "18px 0",
-        }}
-      />
-      <Text
-        style={{
-          margin: "0 6px 12px",
-        }}
-        type="warning"
-      >
-        {messages["annotation.export_no_worker"]}
-        <a href="mailto:hello@webknossos.org">hello@webknossos.org.</a>
-      </Text>
-    </Row>
-  );
-
-  const radioButtonStyle = {
-    marginBottom: 24,
-  };
-
-  const authToken = useFetch(
-    async () => {
-      if (activeUser != null) {
-        return getAuthToken();
-      }
-      return null;
-    },
-    "loading...",
-    [activeUser],
-  );
-
-  const wkInitSnippet = isAnnotation
-    ? getPythonAnnotationDownloadSnippet(authToken, annotation)
-    : getPythonDatasetDownloadSnippet(authToken, dataset);
-
-  const alertTokenIsPrivate = () => {
-    if (authToken) {
-      Toast.warning(
-        "The clipboard contains private data. Do not share this information with anyone you do not trust!",
-      );
-    }
-  };
-
-  const hasSkeleton = annotation.skeleton != null;
-
-  const okText = okTextForTab.get(activeTabKey);
-  const isCurrentlyRunningExportJob =
-    activeTabKey === "export" &&
-    runningExportJobs.some(([key]) => key === exportKey(selectedLayerInfos, mag));
-
-  const isOkButtonDisabled =
-    activeTabKey === "export" &&
-    (!isExportable || isCurrentlyRunningExportJob || isMergerModeEnabled);
-
-  // Will be false if no volumes exist.
-
-  const downloadTab = (
-    <>
-      <Row>
-        {maybeShowWarning()}
-        <Text
-          style={{
-            margin: "0 6px 12px",
-          }}
-        >
-          {!hasVolumes ? "This is a Skeleton-only annotation. " : ""}
-          {!hasSkeleton ? "This is a Volume-only annotation. " : ""}
-          {messages["annotation.download"]}
-        </Text>
-      </Row>
-      <Divider
-        style={{
-          margin: "18px 0",
-        }}
-      >
-        Options
-      </Divider>
-      <Row>
-        <Col
-          span={9}
-          style={{
-            lineHeight: "20px",
-            padding: "5px 12px",
-          }}
-        >
-          Select the data you would like to download.
-          <Hint style={{ marginTop: 12 }}>
-            An NML file will always be included with any download.
-          </Hint>
-        </Col>
-        <Col span={15}>
-          <Radio.Group
-            defaultValue={initialFileFormatToDownload}
-            value={fileFormatToDownload}
-            onChange={(e) => setFileFormatToDownload(e.target.value)}
-            style={{ marginLeft: 16 }}
-          >
-            {hasVolumes ? (
-              <>
-                <Radio value="zarr3" style={radioButtonStyle}>
-                  Include volume annotations as Zarr
-                  <Hint style={{}}>Download a zip folder containing Zarr files.</Hint>
-                </Radio>
-                <Tooltip
-                  title={
-                    isVolumeNDimensional ? "WKW is not supported for n-dimensional volumes." : null
-                  }
-                >
-                  <Radio value="wkw" disabled={isVolumeNDimensional} style={radioButtonStyle}>
-                    Include volume annotations as WKW
-                    <Hint style={{}}>Download a zip folder containing WKW files.</Hint>
-                  </Radio>
-                </Tooltip>
-              </>
-            ) : null}
-            <Radio value="nml" style={radioButtonStyle}>
-              {hasSkeleton ? "Skeleton annotations" : "Meta data"} {hasVolumes ? "only " : ""}
-              as NML
-            </Radio>
-          </Radio.Group>
-        </Col>
-      </Row>
-      <Divider
-        style={{
-          margin: "18px 0",
-        }}
-      />
-      {moreInfoHint}
-    </>
-  );
-
-  const onlyOneMagAvailable = selectedLayerMagInfo.getMagList().length === 1;
-
-  const tiffExportTab = (
-    <>
-      <Row>
-        {maybeShowWarning()}
-        <Text
-          style={{
-            margin: "0 6px 12px",
-          }}
-        >
-          {messages["download.export_as_tiff"]({ typeName })}
-        </Text>
-      </Row>
-      {activeTabKey === "export" &&
-      !dataset.dataStore.jobsSupportedByAvailableWorkers.includes(APIJobCommand.EXPORT_TIFF) ? (
-        workerInfo
-      ) : (
-        <div>
-          <Divider
-            style={{
-              margin: "18px 0",
-            }}
-          >
-            Export format
-          </Divider>
-          <div style={{ display: "flex", justifyContent: "center" }}>
-            <Radio.Group value={exportFormat} onChange={(ev) => setExportFormat(ev.target.value)}>
-              <Radio.Button value={ExportFormat.OME_TIFF}>OME-TIFF</Radio.Button>
-              <Radio.Button value={ExportFormat.TIFF_STACK}>TIFF stack (as .zip)</Radio.Button>
-            </Radio.Group>
-          </div>
-
-          <Divider
-            style={{
-              margin: "18px 0",
-            }}
-          >
-            Layer
-          </Divider>
-          <LayerSelection
-            layers={layers}
-            value={selectedLayerName}
-            onChange={setSelectedLayerName}
-            getReadableNameForLayer={(layer) =>
-              getReadableNameOfVolumeLayer(layer, annotation) || layer.name
-            }
-            style={{ width: "100%" }}
-          />
-
-          <Divider
-            style={{
-              margin: "18px 0",
-            }}
-          >
-            Bounding Box
-          </Divider>
-          <BoundingBoxSelection
-            value={selectedBoundingBoxId}
-            userBoundingBoxes={userBoundingBoxes}
-            setSelectedBoundingBoxId={(boxId: number | null) => {
-              if (boxId != null) {
-                setSelectedBoundingBoxId(boxId);
-              }
-            }}
-            style={{ width: "100%" }}
-          />
-          {boundingBoxCompatibilityAlerts}
-          {(selectedLayerInfos.additionalAxes?.length || 0) > 0 && (
-            <Row>
-              <Divider
-                style={{
-                  margin: "18px 0",
-                }}
-              >
-                Additional Coordinates
-              </Divider>
-              <Text
-                style={{
-                  margin: "0 6px 12px",
-                }}
-              >
-                Your dataset has more than three dimensions. The export will only include the
-                selected bounding box at the current additional dimensions:{" "}
-                {getAdditionalCoordinatesAsString(currentAdditionalCoordinates)}
-              </Text>
-            </Row>
-          )}
-
-          <Divider
-            style={{
-              margin: "18px 0",
-            }}
-          >
-            Mag
-          </Divider>
-          {!onlyOneMagAvailable && (
-            <Row>
-              <Col span={19}>
-                <MagSlider magnificationInfo={selectedLayerMagInfo} value={mag} onChange={setMag} />
-              </Col>
-              <Col
-                span={5}
-                style={{ display: "flex", justifyContent: "flex-end", alignItems: "center" }}
-              >
-                {mag.join("-")}
-              </Col>
-            </Row>
-          )}
-          <Text
-            style={{
-              margin: "0 6px 12px",
-              display: "block",
-            }}
-          >
-            {onlyOneMagAvailable && <div>{mag.join("-")}</div>}
-            Estimated file size:{" "}
-            {estimateFileSize(selectedLayer, mag, selectedBoundingBox.boundingBox, exportFormat)}
-            <br />
-            Magnification: {formatSelectedScale(dataset, mag)}
-          </Text>
-
-          <Divider />
-          <p>
-            Go to the{" "}
-            <a href="/jobs" target="_blank" rel="noreferrer">
-              Jobs Overview Page
-            </a>{" "}
-            to see running exports and to download the results.
-          </p>
-        </div>
-      )}
-      <Divider
-        style={{
-          margin: "18px 0",
-        }}
-      />
-      {moreInfoHint}
-      <Checkbox
-        style={{ position: "absolute", bottom: -38 }}
-        checked={keepWindowOpen}
-        onChange={handleKeepWindowOpenChecked}
-        disabled={activeTabKey === "export" && !features().jobsEnabled}
-      >
-        Keep window open
-      </Checkbox>
-    </>
-  );
-
-  const pythonClientTab = (
-    <>
-      <Row>
-        <Text
-          style={{
-            margin: "0 6px 12px",
-          }}
-        >
-          The following code snippets are suggestions to get you started quickly with the{" "}
-          <a href="https://docs.webknossos.org/webknossos-py/" target="_blank" rel="noreferrer">
-            WEBKNOSSOS Python API
-          </a>
-          . To download and use this {typeName} in your Python project, simply copy and paste the
-          code snippets to your script.
-        </Text>
-      </Row>
-      <Divider
-        style={{
-          margin: "18px 0",
-        }}
-      >
-        Code Snippets
-      </Divider>
-      {maybeShowWarning()}
-      <Paragraph>
-        <CopyableCodeSnippet code="pip install webknossos" />
-        <CopyableCodeSnippet code={wkInitSnippet} onCopy={alertTokenIsPrivate} />
-      </Paragraph>
-      <Divider
-        style={{
-          margin: "18px 0",
-        }}
-      />
-      {moreInfoHint}
-    </>
-  );
-
   const tabs: TabsProps["items"] = [
-    { label: "TIFF Export", key: "export", children: tiffExportTab },
-    { label: "Python Client", key: "python", children: pythonClientTab },
+    {
+      label: "TIFF Export",
+      key: "export",
+      children: (
+        <TiffExportTab
+          dataset={dataset}
+          annotation={annotation}
+          isAnnotation={isAnnotation}
+          initialBoundingBoxId={initialBoundingBoxId}
+        />
+      ),
+    },
+    {
+      label: "Python Client",
+      key: "python",
+      children: (
+        <DownloadPythonTab isAnnotation={isAnnotation} annotation={annotation} dataset={dataset} />
+      ),
+    },
   ];
-  if (isAnnotation) tabs.unshift({ label: "Download", key: "download", children: downloadTab });
+  if (isAnnotation)
+    tabs.unshift({
+      label: "Download",
+      key: "download",
+      children: <DownloadAnnotationTab annotation={annotation} isAnnotation={isAnnotation} />,
+    });
 
   return (
     <Modal
@@ -775,17 +105,18 @@ function _DownloadModalView({
       open={isOpen}
       width={600}
       footer={
-        okText != null ? (
-          <Button
-            key="ok"
-            type="primary"
-            disabled={isOkButtonDisabled}
-            onClick={handleOk}
-            loading={isCurrentlyRunningExportJob}
-          >
-            {okText}
-          </Button>
-        ) : null
+        null
+        // okText != null ? (
+        //   <Button
+        //     key="ok"
+        //     type="primary"
+        //     disabled={isOkButtonDisabled}
+        //     onClick={handleOk}
+        //     loading={isCurrentlyRunningExportJob}
+        //   >
+        //     {okText}
+        //   </Button>
+        //) : null
       }
       onCancel={onClose}
       style={{ overflow: "visible" }}
